requirements.txt
APScheduler==3.6.3
attrs==21.2.0
certifi==2020.12.5
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
